api/requirements.txt
bcrypt==3.1.7
bson==0.5.8
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.20.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
pymongo[tls,srv]==3.8.0
python-dateutil==2.8.0
python-dotenv==0.10.3
six==1.12.0
Werkzeug==0.15.4
